+++ qupath-gui-fx/src/main/java/qupath/lib/gui/QuPathGUI.java
@@ -4,7 +4,7 @@
  * %%
  * Copyright (C) 2014 - 2016 The Queen's University of Belfast, Northern Ireland
  * Contact: IP Management (ipmanagement@qub.ac.uk)
- * Copyright (C) 2018 - 2023 QuPath developers, The University of Edinburgh
+ * Copyright (C) 2018 - 2025 QuPath developers, The University of Edinburgh
  * %%
  * QuPath is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -1110,8 +1110,9 @@ public class QuPathGUI {
 	private void closeAllOpenImagesWithoutPrompts() {
 		for (QuPathViewer v : getAllViewers()) {
 			try {
-				if (v.getImageData() != null)
-					v.getImageData().getServer().close();
+				var imageData = v.getImageData();
+				if (imageData != null)
+					imageData.close();
 			} catch (Exception e2) {
 				logger.warn("Problem closing server", e2);
 			}