+++ qupath-gui-fx/src/main/java/qupath/lib/gui/commands/MiniViewers.java
@@ -4,7 +4,7 @@
  * %%
  * Copyright (C) 2014 - 2016 The Queen's University of Belfast, Northern Ireland
  * Contact: IP Management (ipmanagement@qub.ac.uk)
- * Copyright (C) 2018 - 2022 QuPath developers, The University of Edinburgh
+ * Copyright (C) 2018 - 2025 QuPath developers, The University of Edinburgh
  * %%
  * QuPath is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -25,6 +25,7 @@ package qupath.lib.gui.commands;
 
 import java.awt.AlphaComposite;
 import java.awt.Graphics2D;
+import java.awt.Shape;
 import java.awt.geom.AffineTransform;
 import java.awt.geom.Point2D;
 import java.awt.image.BufferedImage;
@@ -36,6 +37,10 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+
+import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleObjectProperty;
+import javafx.scene.paint.Color;
 import org.controlsfx.control.action.Action;
 import org.controlsfx.control.action.ActionUtils;
 import org.slf4j.Logger;
@@ -88,12 +93,16 @@ import qupath.lib.color.ColorToolsAwt;
 import qupath.lib.common.ColorTools;
 import qupath.lib.common.ThreadTools;
 import qupath.lib.display.ChannelDisplayInfo;
+import qupath.lib.display.DirectServerChannelInfo;
 import qupath.lib.display.ImageDisplay;
 import qupath.lib.gui.actions.ActionTools;
 import qupath.lib.gui.images.stores.AbstractImageRenderer;
 import qupath.lib.gui.images.stores.ImageRenderer;
 import qupath.lib.gui.prefs.PathPrefs;
 import qupath.lib.gui.viewer.QuPathViewer;
+import qupath.lib.gui.viewer.QuPathViewerListener;
+import qupath.lib.images.ImageData;
+import qupath.lib.objects.PathObject;
 import qupath.lib.regions.ImageRegion;
 
 /**
@@ -110,16 +119,45 @@ import qupath.lib.regions.ImageRegion;
 public class MiniViewers {
 	
 	private static final Logger logger = LoggerFactory.getLogger(MiniViewers.class);
-	
-	private static BooleanProperty showAllChannels = PathPrefs.createPersistentPreference("channelViewerAllChannels", false);
+
+	/**
+	 * The type of synchronization to use for mini viewers.
+	 */
+	public enum SyncType {
+		/**
+		 * Center the miniviewer based on the cursor location over the main viewer.
+		 */
+		CURSOR,
+		/**
+		 * Center the miniviewer using the central pixel in the main viewer.
+		 */
+		VIEWER_CENTER,
+		/**
+		 * Center the miniviewer using the last selected object in the main viewer,
+		 * of the main viewer center if no object is selected.
+		 */
+		SELECTED_OBJECT,
+		/**
+		 * Do not synchronize to the main viewer at all.
+		 */
+		NONE
+	}
+
+	private static final BooleanProperty showAllChannels = PathPrefs.createPersistentPreference("channelViewerAllChannels", false);
 
 	/**
 	 * Style binding to use the same background color as the main viewer.
 	 */
-	private static ObservableStringValue style = Bindings.createStringBinding(() -> {
+	private static final ObservableStringValue style = Bindings.createStringBinding(
+			MiniViewers::computeStyle,
+			PathPrefs.viewerBackgroundColorProperty());
+
+
+	private static String computeStyle() {
 		int rgb = PathPrefs.viewerBackgroundColorProperty().get();
 		return String.format("-fx-background-color: rgb(%d, %d, %d)", ColorTools.red(rgb), ColorTools.green(rgb), ColorTools.blue(rgb));
-	}, PathPrefs.viewerBackgroundColorProperty());
+	}
+
 
 	static Stage createDialog(QuPathViewer viewer, boolean channelViewer) {
 		if (channelViewer)
@@ -199,19 +237,26 @@ public class MiniViewers {
 			return false;
 		}
 	}
-	
-	private static boolean isRGBChannel(ChannelDisplayInfo c) {
+
+	/*
+	 * Check whether a channel of an RGB image is a 'direct' channel
+	 * (i.e. the band or a raster, or Red, Green or Blue for an RGB image).
+	 * Reject other 'derived' channels, i.e. those computed by a color transform
+	 * (e.g. color deconvolved, brightness, saturation).
+	 */
+	private static boolean isDirectChannel(ChannelDisplayInfo c) {
+		if (c instanceof DirectServerChannelInfo) {
+			// Fixes https://github.com/qupath/qupath/issues/1948 introduced in v0.6.0.
+			// Note that this would also return channels for non-RGB images.
+			return true;
+		}
 		var method = c.getMethod();
 		if (method == null)
 			return false;
-		switch (method) {
-		case Red:
-		case Green:
-		case Blue:
-			return true;
-		default:
-			return false;
-		}
+        return switch (method) {
+            case Red, Green, Blue -> true;
+            default -> false;
+        };
 	}
 	
 	private static List<ChannelDisplayInfo> getChannels(ImageDisplay display) {
@@ -219,7 +264,9 @@ public class MiniViewers {
 	}
 	
 	private static List<ChannelDisplayInfo> getChannels(ImageDisplay display, boolean allChannels) {
-		var imageData = display == null ? null : display.getImageData();
+		if (display == null)
+			return Collections.emptyList();
+		var imageData = display.getImageData();
 		if (allChannels || imageData == null) {
 			return display.availableChannels();
 		}
@@ -232,7 +279,7 @@ public class MiniViewers {
 			} else {
 				return display.availableChannels()
 						.stream()
-						.filter(MiniViewers::isRGBChannel)
+						.filter(MiniViewers::isDirectChannel)
 						.toList();
 			}			
 		} else {
@@ -240,11 +287,58 @@ public class MiniViewers {
 			var selected = new HashSet<>(display.selectedChannels());
 			return display.availableChannels()
 					.stream()
-					.filter(c -> selected.contains(c))
+					.filter(selected::contains)
 					.toList();
 		}
 	}
-	
+
+	private static Menu createZoomMenu(final MiniViewerManager manager) {
+		List<RadioMenuItem> radioItems = Arrays.asList(
+				ActionUtils.createRadioMenuItem(createDownsampleAction("400 %", manager.downsample, 0.25)),
+				ActionUtils.createRadioMenuItem(createDownsampleAction("200 %", manager.downsample, 0.5)),
+				ActionUtils.createRadioMenuItem(createDownsampleAction("100 %", manager.downsample, 1)),
+				ActionUtils.createRadioMenuItem(createDownsampleAction("50 %", manager.downsample, 2)),
+				ActionUtils.createRadioMenuItem(createDownsampleAction("25 %", manager.downsample, 4)),
+				ActionUtils.createRadioMenuItem(createDownsampleAction("20 %", manager.downsample, 5)),
+				ActionUtils.createRadioMenuItem(createDownsampleAction("10 %", manager.downsample, 10)),
+				ActionUtils.createRadioMenuItem(createDownsampleAction("5 %", manager.downsample, 20)),
+				ActionUtils.createRadioMenuItem(createDownsampleAction("2 %", manager.downsample, 50)),
+				ActionUtils.createRadioMenuItem(createDownsampleAction("1 %", manager.downsample, 100))
+		);
+
+		ToggleGroup group = new ToggleGroup();
+		for (RadioMenuItem item : radioItems) {
+			item.setToggleGroup(group);
+		}
+
+		Menu menuZoom = new Menu("Zoom...");
+		menuZoom.getItems().addAll(radioItems);
+
+		group.selectToggle(radioItems.get(2));
+		return menuZoom;
+	}
+
+	private static Menu createSyncToMenu(final MiniViewerManager manager) {
+		Menu menuSync = new Menu("Sync to...");
+		ToggleGroup groupSync = new ToggleGroup();
+		for (SyncType syncType : SyncType.values()) {
+			var name = switch (syncType) {
+				case CURSOR -> "Cursor";
+				case VIEWER_CENTER -> "Viewer center";
+				case SELECTED_OBJECT -> "Selected object";
+				case NONE -> "Do not sync";
+			};
+			var action = new Action(name, e -> manager.syncType.setValue(syncType));
+			var item = ActionUtils.createRadioMenuItem(action);
+			groupSync.getToggles().add(item);
+			if (syncType == manager.syncType.get()) {
+				groupSync.selectToggle(item);
+				item.setSelected(true);
+			}
+			menuSync.getItems().add(item);
+		}
+		return menuSync;
+	}
 	
 	/**
 	 * Create and install a popup menu in a MiniViewerManager.
@@ -253,30 +347,16 @@ public class MiniViewers {
 	 * @param isChannelViewer 
 	 * @return
 	 */
-	static ContextMenu createPopup(final MiniViewerManager manager, boolean isChannelViewer) {
-		
-		List<RadioMenuItem> radioItems = Arrays.asList(
-				ActionUtils.createRadioMenuItem(createDownsampleAction("400 %", manager.downsample, 0.25)),
-				ActionUtils.createRadioMenuItem(createDownsampleAction("200 %", manager.downsample, 0.5)),
-				ActionUtils.createRadioMenuItem(createDownsampleAction("100 %", manager.downsample, 1)),	
-				ActionUtils.createRadioMenuItem(createDownsampleAction("50 %", manager.downsample, 2)),		
-				ActionUtils.createRadioMenuItem(createDownsampleAction("25 %", manager.downsample, 4)),		
-				ActionUtils.createRadioMenuItem(createDownsampleAction("20 %", manager.downsample, 5)),		
-				ActionUtils.createRadioMenuItem(createDownsampleAction("10 %", manager.downsample, 10)),		
-				ActionUtils.createRadioMenuItem(createDownsampleAction("5 %", manager.downsample, 20)),		
-				ActionUtils.createRadioMenuItem(createDownsampleAction("2 %", manager.downsample, 50)),		
-				ActionUtils.createRadioMenuItem(createDownsampleAction("1 %", manager.downsample, 100))
-		);
-
-		ToggleGroup group = new ToggleGroup();
-		for (RadioMenuItem item : radioItems)
-			item.setToggleGroup(group);
-
-		Menu menuZoom = new Menu("Zoom...");
-		menuZoom.getItems().addAll(radioItems);
+	private static void createPopup(final MiniViewerManager manager, boolean isChannelViewer) {
 
 		ContextMenu popup = new ContextMenu();
-		
+
+		popup.getItems().addAll(
+				createSyncToMenu(manager),
+				createZoomMenu(manager),
+				new SeparatorMenuItem()
+		);
+
 		if (isChannelViewer) {
 			popup.getItems().add(
 					ActionUtils.createCheckMenuItem(ActionTools.createSelectableAction(showAllChannels, "Show all channels"))
@@ -284,22 +364,15 @@ public class MiniViewers {
 		}
 
 		popup.getItems().addAll(
-				ActionUtils.createCheckMenuItem(ActionTools.createSelectableAction(manager.synchronizeToMainViewer, "Synchronize to main viewer")),
-				menuZoom,
-				new SeparatorMenuItem(),
-				ActionUtils.createCheckMenuItem(ActionTools.createSelectableAction(manager.showCursor, "Show cursor")),
 				ActionUtils.createCheckMenuItem(ActionTools.createSelectableAction(manager.showChannelNames, "Show channel names")),
+				ActionUtils.createCheckMenuItem(ActionTools.createSelectableAction(manager.showCursor, "Show cursor")),
 				ActionUtils.createCheckMenuItem(ActionTools.createSelectableAction(manager.showOverlays, "Show overlays"))
 				);
-
-		group.selectToggle(radioItems.get(2));
 				
 		Pane pane = manager.getPane();
 		pane.setOnContextMenuRequested(e -> {
 			popup.show(pane, e.getScreenX(), e.getScreenY());
 		});
-
-		return popup;
 	}
 	
 	
@@ -341,70 +414,59 @@ public class MiniViewers {
 	/**
 	 * A manager for one or more mini-viewers, where the 'more' means a separate viewer per channel.
 	 */
-	public static class MiniViewerManager implements EventHandler<MouseEvent> {
+	public static class MiniViewerManager implements EventHandler<MouseEvent>, QuPathViewerListener {
+
+		private final ObjectProperty<SyncType> syncType = new SimpleObjectProperty<>(SyncType.CURSOR);
+
+		private final GridPane pane = new GridPane();
 		
-		private GridPane pane = new GridPane();
+		private final List<MiniViewer> miniViewers = new ArrayList<>();
 		
-		private List<MiniViewer> miniViewers = new ArrayList<>();
-		
-		private BooleanProperty showChannelNames = new SimpleBooleanProperty(true);
-		private BooleanProperty showCursor = new SimpleBooleanProperty(true);
-		private BooleanProperty showOverlays = new SimpleBooleanProperty(true);
-		private BooleanProperty synchronizeToMainViewer = new SimpleBooleanProperty(true);
-		
+		private final BooleanProperty showChannelNames = new SimpleBooleanProperty(true);
+		private final BooleanProperty showCursor = new SimpleBooleanProperty(true);
+		private final BooleanProperty showOverlays = new SimpleBooleanProperty(true);
+
 		/**
 		 * Track if the shift button is pressed; temporarily suspend synchronization if so.
 		 */
-		private BooleanProperty shiftDown = new SimpleBooleanProperty(false);
+		private final BooleanProperty shiftDown = new SimpleBooleanProperty(false);
 		
-		private DoubleProperty downsample = new SimpleDoubleProperty(1.0);
-		
+		private final DoubleProperty downsample = new SimpleDoubleProperty(1.0);
+
 		private boolean requestUpdate = false;
 		
-		private QuPathViewer mainViewer;
-		private List<ChannelDisplayInfo> channels = new ArrayList<>();
+		private final QuPathViewer mainViewer;
+		private final List<ChannelDisplayInfo> channels = new ArrayList<>();
 		
-		private Point2D centerPosition = new Point2D.Double();
+		private final Point2D centerPosition = new Point2D.Double();
 		private Point2D mousePosition;
 		
-		private ExecutorService pool = Executors.newSingleThreadExecutor(ThreadTools.createThreadFactory("mini-viewer", true));
+		private final ExecutorService pool = Executors.newSingleThreadExecutor(ThreadTools.createThreadFactory("mini-viewer", true));
 		
-		private ChangeListener<Number> changeListener = new ChangeListener<>() {
+		private final ChangeListener<Number> changeListener = this::handleFullUpdateChange;
 
-            @Override
-            public void changed(ObservableValue<? extends Number> observable, Number oldValue, Number newValue) {
-                pool.submit(() -> requestFullUpdate());
-            }
-
-        };
+		private final ChangeListener<Number> fastChangeListener = this::handleFastUpdateChange;
 		
-		private ChangeListener<Number> fastChangeListener = new ChangeListener<>() {
+		private final InvalidationListener invalidationListener = this::handleInvalidation;
 
-            @Override
-            public void changed(ObservableValue<? extends Number> observable, Number oldValue, Number newValue) {
-                pool.submit(() -> requestUpdate());
-            }
+		private void handleFullUpdateChange(ObservableValue<? extends Number> observable, Number oldValue, Number newValue) {
+			pool.submit(this::requestFullUpdate);
+		}
 
-        };
-		
-		private InvalidationListener invalidationListener = new InvalidationListener() {
+		private void handleFastUpdateChange(ObservableValue<? extends Number> observable, Number oldValue, Number newValue) {
+			pool.submit(this::requestUpdate);
+		}
 
-			@Override
-			public void invalidated(Observable observable) {
-				pool.submit(() -> requestUpdate());
-			}
-			
-		};
-		
+		private void handleInvalidation(Observable observable) {
+			pool.submit(this::requestUpdate);
+		}
+
 		/**
 		 * Constructor specifying a primary viewer and number of channels.
 		 * @param mainViewer the viewer that the mini viewers relate to (i.e. tracking the cursor location)
 		 * @param channels the channels to include
-		 * 
-		 * @deprecated use {@link MiniViewers#createManager(QuPathViewer, Collection)} instead.
 		 */
-		@Deprecated
-		public MiniViewerManager(final QuPathViewer mainViewer, final Collection<? extends ChannelDisplayInfo> channels) {
+		private MiniViewerManager(final QuPathViewer mainViewer, final Collection<? extends ChannelDisplayInfo> channels) {
 			this.mainViewer = mainViewer;
 			setChannels(channels);
 
@@ -415,9 +477,10 @@ public class MiniViewers {
 			showCursor.addListener(invalidationListener);
 			showOverlays.addListener(invalidationListener);
 			showChannelNames.addListener(invalidationListener);
-			synchronizeToMainViewer.addListener(invalidationListener);
+			syncType.addListener(invalidationListener);
 			downsample.addListener(invalidationListener);
 			mainViewer.getView().addEventFilter(MouseEvent.MOUSE_MOVED, this);
+			mainViewer.addViewerListener(this);
 			
 			requestFullUpdate();
 		}
@@ -430,14 +493,15 @@ public class MiniViewers {
 			showCursor.removeListener(invalidationListener);
 			showOverlays.removeListener(invalidationListener);
 			showChannelNames.removeListener(invalidationListener);
-			synchronizeToMainViewer.removeListener(invalidationListener);
+			syncType.removeListener(invalidationListener);
 			downsample.removeListener(invalidationListener);
 			mainViewer.getView().removeEventFilter(MouseEvent.MOUSE_MOVED, this);
+			mainViewer.removeViewerListener(this);
 		}
 		
 		void setChannels(Collection<? extends ChannelDisplayInfo> channels) {
 			
-			miniViewers.stream().forEach(v -> v.close());
+			miniViewers.forEach(MiniViewer::close);
 			miniViewers.clear();
 			
 			int nChannels = channels.size();
@@ -503,7 +567,7 @@ public class MiniViewers {
 			Label label = new Label();
 			label.textProperty().bind(canvas.nameBinding);
 			label.setStyle("-fx-background-color: rgba(0, 0, 0, 0.6)");
-			label.setTextFill(javafx.scene.paint.Color.WHITE);
+			label.setTextFill(Color.WHITE);
 			label.setAlignment(Pos.CENTER);
 			double height = 50;
 			label.setPrefHeight(height);
@@ -561,31 +625,81 @@ public class MiniViewers {
 				return;
 			// TODO: Look to improve the performance of this & discard requests appropriately
 			requestUpdate = true;
-			Platform.runLater(() -> updateViewers());
-//			updateViewers();
+			Platform.runLater(this::updateViewers);
 		}
-		
+
+		@Override
+		public void imageDataChanged(QuPathViewer viewer, ImageData<BufferedImage> imageDataOld, ImageData<BufferedImage> imageDataNew) {}
+
+		@Override
+		public void visibleRegionChanged(QuPathViewer viewer, Shape shape) {
+			if (syncType.get() != SyncType.CURSOR) {
+				// Update for center OR selected object, because if no object is selected then we use center by default
+				requestUpdate();
+			}
+		}
+
+		@Override
+		public void selectedObjectChanged(QuPathViewer viewer, PathObject pathObjectSelected) {
+			if (syncType.get() == SyncType.SELECTED_OBJECT)
+				requestUpdate();
+		}
+
+		@Override
+		public void viewerClosed(QuPathViewer viewer) {}
+
 		void updateViewers() {
 			if (!requestUpdate)
 				return;
-//			if (!Platform.isFxApplicationThread()) {
-//				Platform.runLater(() -> updateViewers());
-//				return;
-//			}
-			if (this.pane != null && this.pane.isVisible()) {
+			if (pane.isVisible()) {
 				// Update mouse position if we are synchronizing
 				mousePosition = mainViewer.getMousePosition();
 				if (mousePosition != null) {
 					mainViewer.componentPointToImagePoint(mousePosition, mousePosition, false);
-					if (synchronizeToMainViewer.get() && !shiftDown.get())
-						centerPosition.setLocation(mousePosition);
 				}
-	
+				updatePosition();
 				// Repaint all viewers
 				for (MiniViewer viewer : miniViewers)
 					viewer.repaint();
 			}
 			requestUpdate = false;
+		}
+
+		/*
+		 * Update the position of the center pixel used by the mini viewers.
+		 */
+		private void updatePosition() {
+			var sync = syncType.get();
+			if (sync == null)
+				return;
+			switch (sync) {
+				case CURSOR -> updatePositionFromCursor();
+				case VIEWER_CENTER -> updatePositionFromCenter();
+				case SELECTED_OBJECT -> updatePositionFromSelected();
+			};
+		}
+
+		private void updatePositionFromCursor() {
+			if (mousePosition != null && !shiftDown.get()) {
+				centerPosition.setLocation(mousePosition);
+			}
+		}
+
+		private void updatePositionFromCenter() {
+			centerPosition.setLocation(mainViewer.getCenterPixelX(), mainViewer.getCenterPixelY());
+		}
+
+		private void updatePositionFromSelected() {
+			var selected = mainViewer.getSelectedObject();
+			var roi = selected == null ? null : selected.getROI();
+			if (roi != null) {
+				centerPosition.setLocation(
+						roi.getBoundsX() + roi.getBoundsWidth()/2.0,
+						roi.getBoundsY() + roi.getBoundsHeight()/2.0
+				);
+			} else {
+				updatePositionFromCenter();
+			}
 		}
 		
 		
@@ -599,15 +713,15 @@ public class MiniViewers {
 		
 		class MiniViewer extends Canvas {
 			
-			private ImageRenderer renderer;
+			private final ImageRenderer renderer;
 						
-			private StringProperty nameBinding = new SimpleStringProperty();
+			private final StringProperty nameBinding = new SimpleStringProperty();
 			private BufferedImage imgRGB;
 			private BufferedImage img;
 			private WritableImage imgFX;
 			
 			private Point2D localCursorPosition = null;
-			private AffineTransform transform = new AffineTransform();
+			private final AffineTransform transform = new AffineTransform();
 
 			
 			public MiniViewer(ImageRenderer renderer) {
@@ -618,9 +732,8 @@ public class MiniViewers {
 				
 				// Create binding to indicate the current channel name
 				nameBinding.bind(Bindings.createStringBinding(() -> {
-						if (renderer instanceof ImageDisplaySingleChannelRenderer) {
-							ImageDisplaySingleChannelRenderer channelRenderer = (ImageDisplaySingleChannelRenderer)renderer;
-							int channel = channelRenderer.channel;
+						if (renderer instanceof ImageDisplaySingleChannelRenderer channelRenderer) {
+                            int channel = channelRenderer.channel;
 							if (channel < 0 || channel >= channels.size())
 								return null;
 							return channels.get(channel).getName();
@@ -756,14 +869,14 @@ public class MiniViewers {
 				
 				if (showCursor.get() && localCursorPosition != null) {
 					context.setLineWidth(4);
-					context.setStroke(javafx.scene.paint.Color.BLACK);
+					context.setStroke(Color.BLACK);
 					double len = 4.0;
 					double x = localCursorPosition.getX();
 					double y = localCursorPosition.getY();
 					context.strokeLine(x, y-len, x, y+len);
 					context.strokeLine(x-len, y, x+len, y);
 					context.setLineWidth(2);
-					context.setStroke(javafx.scene.paint.Color.WHITE);
+					context.setStroke(Color.WHITE);
 					context.strokeLine(x, y-len, x, y+len);
 					context.strokeLine(x-len, y, x+len, y);
 				}